requirements.txt
discord.py==2.5.2
Flask==3.1.1
